Add test layers
